+++ src/runtime/index.js
@@ -7,6 +7,7 @@ export { Hook } from './hook';
 export { Callback } from './callback';
 export { Timer } from './timer';
 export { CommandInterpreter } from './command_interpreter';
+export { mountExospine } from './exospine';
 export { SseConnector } from './sse_connector';
 export { CommandClient } from './command_client';
 export { useNodeState, useNodeFill } from './react';